+++ examples/example/vue3/src/examples/datepicker/03/App.vue
@@ -8,8 +8,8 @@ const value = ref<Date | null>(null)
 <template>
   <DatePicker
     v-model="value"
-    :min-date="new Date('2026-01-01')"
-    :max-date="new Date('2026-12-31')"
+    min-date="2026-01-01"
+    max-date="2026-12-31"
     placeholder="仅可选择 2026 年日期"
     class="w-full max-w-[280px]" />
 </template>